Minor fixes for Linux/GCC
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@1201 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation/TurnManager.cpp
@@ -70,7 +70,7 @@ void CTurnManager::SendBatch(uint batch)
 		SendMessage(it->m_pMessage, it->m_ClientMask);
 		++it;
 	}
-	SendMessage(new CEndCommandBatch(), -1);
+	SendMessage(new CEndCommandBatch(), (uint)-1);
 }
 
 void CTurnManager::SendMessage(CNetMessage *pMsg, uint clientMask)
